Prevent synth note hangs
It was possible to time a stop after the time a note
was sent to midi and before the time the note actually was
scheduled for.  This fix accounts for this case and also the
case where a note was scheduled after isPlaying was already set to
false but before the check for isPlaying.